Added JoinLeaveEventListener which handles new join events Added CommandList which contains all command objects
