+++ src/command/CommandList.java
@@ -1,0 +1,40 @@
+package command;
+
+import java.util.ArrayList;
+
+/**
+ * Container class for a list of all the commands
+ * available in the server.
+ */
+public class CommandList {
+
+    private static ArrayList<Command> commands;
+
+
+    /**
+     * Initializes the ArrayList of commands and
+     * populates it with every possible command.
+     */
+    static {
+        commands = new ArrayList<>();
+        addAllCommands();
+    }
+
+
+    /**
+     * Commands list getter.
+     *
+     * @return the ArrayList of all commands
+     */
+    public static ArrayList<Command> getCommands() {
+        return commands;
+    }
+
+
+    /**
+     * Populates the commands list with every command.
+     */
+    private static void addAllCommands() {
+        commands.add(new Ping());
+    }
+}
